requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.65.0
opencv-python>=4.5.0
lxml==5.1.0 